[US164Task178] Modified after PR Review
